CUDA/OpenCL 依赖改为 SYSTEM include，隔离供应商头文件诊断。 严格告警改为仅应用于 executor 库目标，避免将库 API 规则误施加到独立测试入口。 修复库内的真实类型转换、未使用变量、声明缺失、noexcept 与 C++20 volatile 问题。 -Wstrict-overflow 保留 GCC 标准的 level 1 检查；移除了无法定位源码的 level 5 模板优化噪声，没有使用 -Wno-* 压制告警。

+++ src/executor/gpu/device_query.cpp
@@ -33,10 +33,10 @@ std::vector<GpuDeviceInfo> enumerate_cuda_devices() {
         info.vendor = "NVIDIA";
         info.compute_capability_major = prop.major;
         info.compute_capability_minor = prop.minor;
-        info.max_threads_per_block = prop.maxThreadsPerBlock;
-        info.max_blocks_per_grid[0] = prop.maxGridSize[0];
-        info.max_blocks_per_grid[1] = prop.maxGridSize[1];
-        info.max_blocks_per_grid[2] = prop.maxGridSize[2];
+        info.max_threads_per_block = static_cast<size_t>(prop.maxThreadsPerBlock);
+        info.max_blocks_per_grid[0] = static_cast<size_t>(prop.maxGridSize[0]);
+        info.max_blocks_per_grid[1] = static_cast<size_t>(prop.maxGridSize[1]);
+        info.max_blocks_per_grid[2] = static_cast<size_t>(prop.maxGridSize[2]);
 
         // 获取内存信息
         size_t free_mem = 0, total_mem = 0;
